Init channel instance as soon as all positions are gathered
Instead of sending a special message, oddly named TEST_MATRIX, to
initialize the channel, this initialization will now be conducted upon
receipt of the last set of Mobile Station Positions.

+++ src/BsChannel.h
@@ -33,7 +33,7 @@ class BsChannel : public cSimpleModule  {
         bool useSimpleChannelCalc;
         int simpleChannelCalcNops;
         int bsId;
-        int init_counter;
+        size_t init_counter;
         int numberOfVR;
         Position bsPosition;
         double packetLoss;
